% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot.R
\name{my_plot}
\alias{my_plot}
\title{my_plot}
\usage{
my_plot(data, x, y, title)
}
\arguments{
\item{data}{\code{data.frame} containing data}

\item{x}{column containing x-axis data}

\item{y}{column containing y-axis data}

\item{title}{\code{string} containing plot title}
}
\value{
Returns a \code{ggplot} object
}
\description{
A very simple plotting function
}
